requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.8.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
DateTime==4.3
dnspython==1.16.0
future==0.18.2
idna==2.8
numpy==1.18.1
pandas==0.25.3
pycparser==2.19
pymongo==3.10.0
python-dateutil==2.8.1
python-telegram-bot==12.3.0
pytz==2019.3
requests==2.22.0
selenium==3.141.0
six==1.13.0
soupsieve==1.9.5
tornado==6.0.3
tzlocal==2.0.0
urllib3==1.25.7
zope.interface==4.7.1
